+++ MainGUI/Resources/change_log.md
@@ -5,7 +5,7 @@ Modnix Changelog
 * New: See console log and changelog in Modnix.
 * New: Loader log now auto-refresh, same for new console log.
 * New: Loader log now can be filtered, same for new console log.
-* New: Mod that logs an exception as error will be detected and show a warning.
+* New: Mod that logs an error or warning will be detected and show a warning.
 * Fix: Embedded mod_info with BOM can now be correctly read.
 * Fix: Mods with multiple non-static initialiser classes are now called with the correct classes.
 * Fix: mods_root added as a reserved id. It is used by mod api path action.